Small improvement.
